requirements.txt
# ============================================================
# ICDAR 2026 DocVQA — Python Dependencies
# ============================================================
# Python : 3.10.x
# CUDA   : 12.1
#
# Install:
#   pip install --upgrade pip
#   pip install -r requirements.txt
# ============================================================

# ── PyTorch + CUDA 12.1 ──
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.3.0+cu121
torchvision==0.18.0+cu121

# ── Vision-Language Models ──
transformers==4.49.0
accelerate==0.33.0
bitsandbytes==0.43.3
peft==0.13.2
qwen-vl-utils==0.0.1

# ── PaddleOCR (Parser) ──
paddlepaddle-gpu==2.6.2
paddleocr==2.8.1

# ── Alternative Parser ──
docling==2.5.0

# ── Dataset + HuggingFace ──
datasets==2.21.0
huggingface_hub==0.24.0
Pillow==10.4.0
sentencepiece==0.2.0

# ── Evaluation ──
numpy==1.26.0
rapidfuzz==3.9.0
jiwer==3.0.4
python-Levenshtein==0.25.1
python-dateutil==2.9.0

# ── Utilities ──
pandas==2.2.0
tqdm==4.66.0
pyyaml==6.0.1
python-dotenv==1.0.1
matplotlib==3.9.0
seaborn==0.13.0

# ── Experiment Tracking (optional) ──
wandb==0.18.0

# ── Testing ──
pytest==8.2.0
pytest-cov==5.0.0
